requirements.txt
Django>=2.0
requests>=2.18
django-mathfilters>=0.4.0
gunicorn>=19.7
dj_database_url>=0.4.2
psycopg2-binary==2.7.5
whitenoise==3.3.1
numpy==1.14.3
scipy==1.2.0
Pillow==6.2.0

discord.py==1.2.4
